@@ -9,21 +9,22 @@
 		"check:watch": "svelte-check --tsconfig ./tsconfig.json --watch"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^1.0.0-next.87",
-		"@sveltejs/kit": "^1.0.0-next.538",
+		"@sveltejs/adapter-auto": "^3.0.0",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@types/domutils": "^1.7.4",
 		"@types/htmlparser2": "^3.10.3",
 		"@types/sharp": "^0.30.4",
 		"autoprefixer": "^10.4.7",
 		"postcss": "^8.4.14",
 		"sass": "^1.52.3",
-		"svelte": "^3.44.0",
-		"svelte-check": "^2.7.1",
-		"svelte-preprocess": "^4.10.7",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
+		"svelte-preprocess": "^6.0.0",
 		"tailwindcss": "^3.1.3",
 		"tslib": "^2.3.1",
-		"typescript": "^4.7.2",
-		"vite": "^3.1.0"
+		"typescript": "^5.5.0",
+		"vite": "^5.4.4"
 	},
 	"type": "module",
 	"dependencies": {
