@@ -4,19 +4,28 @@
   import IoIosSpeedometer from 'svelte-icons/io/IoIosSpeedometer.svelte';
   import SlotCardLowerRowFact from "./SlotCardLowerRowFact.svelte";
 
-  export let lastLap: Slot['lastLap'];
-  export let fastestLap: Slot['fastestLap'];
-  export let remainingGas: Slot['remainingGas'];
-  export let inPit: Slot['inPit'];
+  interface Props {
+    lastLap: Slot['lastLap'];
+    fastestLap: Slot['fastestLap'];
+    remainingGas: Slot['remainingGas'];
+    inPit: Slot['inPit'];
+  }
 
-  let width: number;
+  let {
+    lastLap,
+    fastestLap,
+    remainingGas,
+    inPit
+  }: Props = $props();
 
-  $: isDesktop = !width || width >= 390;
+  let width: number = $state();
 
-  $: gasGreen = remainingGas > 0.8;
-  $: gasRed = remainingGas < 0.3;
-  $: gasYellow = !gasGreen && !gasRed;
-  $: gasPulsing = remainingGas < 0.2;
+  let isDesktop = $derived(!width || width >= 390);
+
+  let gasGreen = $derived(remainingGas > 0.8);
+  let gasRed = $derived(remainingGas < 0.3);
+  let gasYellow = $derived(!gasGreen && !gasRed);
+  let gasPulsing = $derived(remainingGas < 0.2);
 </script>
 
 
@@ -28,8 +37,12 @@
 >
     <div class="w-[6.4rem]" class:w-[5rem]={!isDesktop}>
         <SlotCardLowerRowFact>
-            <IoMdStopwatch slot="icon"/>
-            <span slot="title">Letzte {isDesktop ? 'Runde' : ''}</span>
+            {#snippet icon()}
+            <IoMdStopwatch />
+          {/snippet}
+            {#snippet title()}
+            <span >Letzte {isDesktop ? 'Runde' : ''}</span>
+          {/snippet}
 
             {#if lastLap?.time}
                 {lastLap.time}s
@@ -40,8 +53,12 @@
     </div>
     <div class="w-[7.9rem]" class:w-[5.5rem]={!isDesktop}>
         <SlotCardLowerRowFact>
-            <IoMdStopwatch slot="icon"/>
-            <span slot="title">Schnellste {isDesktop ? 'Runde' : ''}</span>
+            {#snippet icon()}
+            <IoMdStopwatch />
+          {/snippet}
+            {#snippet title()}
+            <span >Schnellste {isDesktop ? 'Runde' : ''}</span>
+          {/snippet}
 
             {#if fastestLap?.time}
                 {fastestLap.time}s
@@ -52,19 +69,25 @@
     </div>
     <div class="w-[6.2rem]">
         <SlotCardLowerRowFact>
-            <IoIosSpeedometer slot="icon"/>
-            <span slot="title">Tankstand</span>
-            <div slot="indicator" class="w-full h-full">
-                <div class="w-full h-full shadow-sm transition-colors"
-                     class:bg-green-600={gasGreen}
-                     class:shadow-green-600={gasGreen}
-                     class:bg-yellow-400={gasYellow}
-                     class:shadow-yellow-400={gasYellow}
-                     class:bg-red-600={gasRed}
-                     class:shadow-red-600={gasRed}
-                     class:animate-pulse={gasPulsing}
-                ></div>
-            </div>
+            {#snippet icon()}
+            <IoIosSpeedometer />
+          {/snippet}
+            {#snippet title()}
+            <span >Tankstand</span>
+          {/snippet}
+            {#snippet indicator()}
+            <div  class="w-full h-full">
+                  <div class="w-full h-full shadow-sm transition-colors"
+                       class:bg-green-600={gasGreen}
+                       class:shadow-green-600={gasGreen}
+                       class:bg-yellow-400={gasYellow}
+                       class:shadow-yellow-400={gasYellow}
+                       class:bg-red-600={gasRed}
+                       class:shadow-red-600={gasRed}
+                       class:animate-pulse={gasPulsing}
+                  ></div>
+              </div>
+          {/snippet}
             {(remainingGas * 100).toFixed(0)}%
         </SlotCardLowerRowFact>
     </div>
