@@ -1,22 +1,38 @@
+<script lang="ts">
+    interface Props {
+        icon?: import('svelte').Snippet;
+        indicator?: import('svelte').Snippet;
+        title?: import('svelte').Snippet;
+        children?: import('svelte').Snippet;
+    }
+
+    let {
+        icon,
+        indicator,
+        title,
+        children
+    }: Props = $props();
+</script>
+
 <div class="flex items-center gap-x-1.5">
-    {#if $$slots.icon}
+    {#if icon}
         <div class="w-5 h-5 text-neutral-400">
-            <slot name="icon"/>
+            {@render icon?.()}
         </div>
     {/if}
-    {#if $$slots.indicator}
+    {#if indicator}
         <div class="h-7 w-1">
-            <slot name="indicator"/>
+            {@render indicator?.()}
         </div>
     {/if}
     <div>
-        {#if $$slots.title}
+        {#if title}
             <div class="text-xs text-neutral-400 font-normal overflow-hidden text-ellipsis whitespace-nowrap">
-                <slot name="title"/>
+                {@render title?.()}
             </div>
         {/if}
         <div class="text-sm font-normal">
-            <slot/>
+            {@render children?.()}
         </div>
     </div>
 </div>
