@@ -1,9 +1,14 @@
 <script lang="ts">
   import Card from "$lib/components/Card.svelte";
+  interface Props {
+    children?: import('svelte').Snippet;
+  }
+
+  let { children }: Props = $props();
 </script>
 
 <Card>
     <div class="p-4">
-        <slot />
+        {@render children?.()}
     </div>
 </Card>
