@@ -1,6 +1,17 @@
 <script lang="ts">
-    export let disabled = false;
-    export let isLast = false;
+    interface Props {
+        disabled?: boolean;
+        isLast?: boolean;
+        children?: import('svelte').Snippet;
+        hint?: import('svelte').Snippet;
+    }
+
+    let {
+        disabled = false,
+        isLast = false,
+        children,
+        hint
+    }: Props = $props();
 </script>
 <div class="px-5 py-5 border-neutral-200 dark:border-neutral-700 font-normal transition-opacity"
      class:cursor-not-allowed={disabled}
@@ -9,11 +20,11 @@
 >
     <div class="flex flex-col gap-y-0.5"
          class:pointer-events-none={disabled}>
-        <slot />
+        {@render children?.()}
 
-        {#if $$slots.hint}
+        {#if hint}
             <div class="text-neutral-400 text-sm">
-                <slot name="hint" />
+                {@render hint?.()}
             </div>
         {/if}
     </div>
