@@ -5,17 +5,17 @@ import type {ApiData} from '$lib/models/api-data';
 
 export const load: PageLoad = async ({params: {sessionName, slotId}, parent}) => {
   if (!sessionName || !slotId) {
-    throw error(400, 'Session nicht gefunden 🤷');
+    error(400, 'Session nicht gefunden 🤷');
   }
 
   const data = await parent() as ApiData<Race>;
 
   if (!data?.data?.slots) {
-    throw error(404, 'Session nicht gefunden 🤷');
+    error(404, 'Session nicht gefunden 🤷');
   }
 
   if (!data.data.slots.find(slot => slot.id === slotId)) {
-    throw error(404, 'Slot nicht gefunden 🤷')
+    error(404, 'Slot nicht gefunden 🤷');
   }
 
   return data;
